requirements.txt
ffmpeg==1.4
nlpcloud==1.1.46
openai-whisper==20240930
pocketsphinx==5.0.3
PyAudio==0.2.14
pydub==0.25.1
sounddevice==0.5.1
SpeechRecognition==3.11.0
torch==1.13.0
python-dotenv==1.0.1
SinaTools==0.1.36
pandas==2.2.3
google-api-python-client==2.154.0
